+++ src/resonaate/common/logger.py
@@ -3,11 +3,9 @@
 import logging
 import sys
 from datetime import datetime
+from logging.handlers import RotatingFileHandler
 from os import makedirs
 from os.path import exists, join
-
-# Third Party Imports
-from concurrent_log_handler import ConcurrentRotatingFileHandler as FileHandler
 
 # Local Imports
 from .behavioral_config import BehavioralConfig
@@ -16,7 +14,6 @@ from .behavioral_config import BehavioralConfig
 class Logger:
     """Extended logger wraps the standard Python logging package.
 
-    It utilizes the `concurrent_log_handler` package to avoid dropping log files.
     It also creates a standard file name and log format for any log files that are saved.
     """
 
@@ -58,7 +55,7 @@ class Logger:
                 self.filename = join(path, log_name)
 
                 # Create the file handler based on the file name
-                handler = FileHandler(
+                handler = RotatingFileHandler(
                     self.filename,
                     maxBytes=BehavioralConfig.getConfig().logging.MaxFileSize,
                     backupCount=BehavioralConfig.getConfig().logging.MaxFileCount,